+++ dffml/base.py
@@ -25,7 +25,7 @@ except ImportError:
         return getattr(t, "__args__", None)
 
 
-from .util.cli.arg import Arg
+from .util.cli.plugin import Plugin
 from .util.data import (
     traverse_config_set,
     traverse_config_get,
@@ -68,7 +68,7 @@ def list_action(list_cls):
     return LIST_ACTIONS[list_cls]
 
 
-class MissingArg(Exception):
+class MissingPlugin(Exception):
     """
     Raised when a BaseConfigurable is missing an argument from the args dict it
     created with args(). If this exception is raised then the config() method is
@@ -118,61 +118,61 @@ class BaseConfig(object):
         return repr(self)
 
 
-def mkarg(field):
-    arg = Arg(type=field.type)
+def mkplugin(field):
+    plugin = Plugin(type=field.type)
     # HACK For detecting dataclasses._MISSING_TYPE
     if "dataclasses._MISSING_TYPE" not in repr(field.default):
-        arg["default"] = field.default
+        plugin["default"] = field.default
     if "dataclasses._MISSING_TYPE" not in repr(field.default_factory):
-        arg["default"] = field.default_factory()
+        plugin["default"] = field.default_factory()
     if field.type == bool:
-        arg["action"] = "store_true"
+        plugin["action"] = "store_true"
     elif inspect.isclass(field.type):
         if issubclass(field.type, (list, collections.UserList)):
-            arg["nargs"] = "+"
+            plugin["nargs"] = "+"
             if not hasattr(field.type, "SINGLETON"):
                 raise AttributeError(
                     f"{field.type.__qualname__} missing attribute SINGLETON"
                 )
-            arg["action"] = list_action(field.type)
-            arg["type"] = field.type.SINGLETON
-        if hasattr(arg["type"], "load"):
+            plugin["action"] = list_action(field.type)
+            plugin["type"] = field.type.SINGLETON
+        if hasattr(plugin["type"], "load"):
             # TODO (python3.8) Use Protocol
-            arg["type"] = arg["type"].load
+            plugin["type"] = plugin["type"].load
     elif get_origin(field.type) is list:
-        arg["type"] = get_args(field.type)[0]
-        arg["nargs"] = "+"
+        plugin["type"] = get_args(field.type)[0]
+        plugin["nargs"] = "+"
     if "description" in field.metadata:
-        arg["help"] = field.metadata["description"]
-    return arg
+        plugin["help"] = field.metadata["description"]
+    return plugin
 
 
-def convert_value(arg, value):
+def convert_value(plugin, value):
     if value is None:
         # Return default if not found and available
-        if "default" in arg:
-            return copy.deepcopy(arg["default"])
+        if "default" in plugin:
+            return copy.deepcopy(plugin["default"])
         raise MissingConfig
 
-    if not "nargs" in arg and isinstance(value, list):
+    if not "nargs" in plugin and isinstance(value, list):
         value = value[0]
-    if "type" in arg:
-        type_cls = arg["type"]
+    if "type" in plugin:
+        type_cls = plugin["type"]
         if type_cls == Type:
             type_cls = type_lookup
         # TODO This is a oversimplification of argparse's nargs
-        if "nargs" in arg:
+        if "nargs" in plugin:
             value = list(map(type_cls, value))
         else:
             value = type_cls(value)
-    if "action" in arg:
-        if isinstance(arg["action"], str):
+    if "action" in plugin:
+        if isinstance(plugin["action"], str):
             # HACK This accesses _pop_action_class from ArgumentParser
             # which is prefaced with an underscore indicating it not an API
             # we can rely on
-            arg["action"] = ARGP._pop_action_class(arg)
+            plugin["action"] = ARGP._pop_action_class(plugin)
         namespace = ConfigurableParsingNamespace()
-        action = arg["action"](dest="dest", option_strings="")
+        action = plugin["action"](dest="dest", option_strings="")
         action(None, namespace, value)
         value = namespace.dest
     return value
@@ -180,7 +180,7 @@ def convert_value(arg, value):
 
 def is_config_dict(value):
     return bool(
-        "arg" in value
+        "plugin" in value
         and "config" in value
         and isinstance(value["config"], dict)
     )
@@ -192,19 +192,19 @@ def _fromdict(cls, **kwargs):
             value = kwargs[field.name]
             config = {}
             if is_config_dict(value):
-                value, config = value["arg"], value["config"]
-            value = convert_value(mkarg(field), value)
+                value, config = value["plugin"], value["config"]
+            value = convert_value(mkplugin(field), value)
             if inspect.isclass(value) and issubclass(value, BaseConfigurable):
                 # TODO This probably isn't 100% correct. Figure out what we need
                 # to do with nested configs.
                 value = value.withconfig(
                     {
                         value.ENTRY_POINT_NAME[-1]: {
-                            "arg": None,
+                            "plugin": None,
                             "config": {
                                 key: value
                                 if is_config_dict(value)
-                                else {"arg": value, "config": {}}
+                                else {"plugin": value, "config": {}}
                                 for key, value in config.items()
                             },
                         }
@@ -402,15 +402,15 @@ class BaseConfigurable(metaclass=BaseConfigurableMetaClass):
         label_above = cls.add_label(*above) + list(path)
         no_label_above = cls.add_label(*above)[:-1] + list(path)
 
-        arg = None
+        plugin = None
         try:
-            arg = traverse_config_get(args, *args_above)
+            plugin = traverse_config_get(args, *args_above)
         except KeyError as error:
             pass
 
-        if arg is None:
-            raise MissingArg(
-                "Arg %r missing from %s%s%s"
+        if plugin is None:
+            raise MissingPlugin(
+                "Plugin %r missing from %s%s%s"
                 % (
                     args_above[-1],
                     cls.__qualname__,
@@ -430,7 +430,7 @@ class BaseConfigurable(metaclass=BaseConfigurableMetaClass):
                 value = traverse_config_get(config, *no_label_above)
 
         try:
-            return convert_value(arg, value)
+            return convert_value(plugin, value)
         except MissingConfig as error:
             error.args = (
                 (
@@ -445,7 +445,7 @@ class BaseConfigurable(metaclass=BaseConfigurableMetaClass):
             raise
 
     @classmethod
-    def args(cls, args, *above) -> Dict[str, Arg]:
+    def args(cls, args, *above) -> Dict[str, Plugin]:
         """
         Return a dict containing arguments required for this class
         """
@@ -454,7 +454,7 @@ class BaseConfigurable(metaclass=BaseConfigurableMetaClass):
                 f"{cls.__qualname__} requires CONFIG property or implementation of args() classmethod"
             )
         for field in dataclasses.fields(cls.CONFIG):
-            cls.config_set(args, above, field.name, mkarg(field))
+            cls.config_set(args, above, field.name, mkplugin(field))
         return args
 
     @classmethod