+++ dffml/service/dev.py
@@ -21,7 +21,7 @@ from ..base import BaseConfig
 from ..util.os import chdir, MODE_BITS_SECURE
 from ..version import VERSION
 from ..util.skel import Skel, SkelTemplateConfig
-from ..util.cli.arg import Arg
+from ..util.cli.plugin import Plugin
 from ..util.cli.cmd import CMD
 from ..util.entrypoint import load
 from ..base import MissingConfig
@@ -68,37 +68,39 @@ def create_from_skel(name):
 
         skel = Skel()
 
-        arg_user = Arg(
+        plugin_user = Plugin(
             "-user",
             help=f"Your username (default: {USER})",
             default=USER,
             required=False,
         )
-        arg_name = Arg(
+        plugin_name = Plugin(
             "-name",
             help=f"Your name (default: {NAME})",
             default=NAME,
             required=False,
         )
-        arg_email = Arg(
+        plugin_email = Plugin(
             "-email",
             help=f"Your email (default: {EMAIL})",
             default=EMAIL,
             required=False,
         )
-        arg_description = Arg(
+        plugin_description = Plugin(
             "-description",
             help=f"Description of python package (default: DFFML {name} {{package name}})",
             default=None,
             required=False,
         )
-        arg_target = Arg(
+        plugin_target = Plugin(
             "-target",
             help=f"Directory to put code in (default: same as package name)",
             default=None,
             required=False,
         )
-        arg_package = Arg("package", help="Name of python package to create")
+        plugin_package = Plugin(
+            "package", help="Name of python package to create"
+        )
 
         async def run(self):
             # Set description if None
@@ -161,7 +163,7 @@ class Run(CMD):
     Run a single operation
     """
 
-    arg_operation = Arg("operation", help="Python path to operation")
+    plugin_operation = Plugin("operation", help="Python path to operation")
 
     async def run(self):
         # Push current directory into front of path so we can run things
@@ -182,7 +184,7 @@ class Run(CMD):
             value = traverse_config_get(self.extra_config, key)
         except KeyError as error:
             raise MissingConfig("%s missing %s" % (op.name, key))
-        # TODO Argparse nargs and Arg and primitives need to be unified
+        # TODO Argparse nargs and Plugin and primitives need to be unified
         if "Dict" in definition.primitive:
             # TODO handle Dict / spec completely
             self.logger.critical(
@@ -256,7 +258,7 @@ class Run(CMD):
 
 class ListEntrypoints(CMD):
 
-    arg_entrypoint = Arg(
+    plugin_entrypoint = Plugin(
         "entrypoint", help="Entrypoint to list, example: dffml.model"
     )
 
@@ -272,20 +274,20 @@ class Entrypoints(CMD):
 
 class Export(CMD):
 
-    arg_config = Arg(
+    plugin_config = Plugin(
         "-config",
         help="ConfigLoader to use",
         type=BaseConfigLoader.load,
         default=JSONConfigLoader,
     )
-    arg_not_linked = Arg(
+    plugin_not_linked = Plugin(
         "-not-linked",
         dest="not_linked",
         help="Do not export dataflows as linked",
         default=False,
         action="store_true",
     )
-    arg_export = Arg("export", help="Python path to object to export")
+    plugin_export = Plugin("export", help="Python path to object to export")
 
     async def run(self):
         async with self.config(BaseConfig()) as configloader:
@@ -315,7 +317,7 @@ class Install(CMD):
     Uninstall production packages and install dffml in development mode.
     """
 
-    arg_user = Arg(
+    plugin_user = Plugin(
         "-user", "Preform user install", default=False, action="store_true"
     )
 
@@ -350,8 +352,8 @@ class SetupPyKWArg(CMD):
     Get a keyword argument from a call to setup in a setup.py file.
     """
 
-    arg_kwarg = Arg("kwarg", help="Keyword argument to write to stdout")
-    arg_setup_filepath = Arg("setup_filepath", help="Path to setup.py")
+    plugin_kwarg = Plugin("kwarg", help="Keyword argument to write to stdout")
+    plugin_setup_filepath = Plugin("setup_filepath", help="Path to setup.py")
 
     @staticmethod
     def get_kwargs(setup_filepath: str):
@@ -393,7 +395,7 @@ class Release(CMD):
     Release a package (if not already released)
     """
 
-    arg_package = Arg(
+    plugin_package = Plugin(
         "package", help="Relative path to package to release", type=Path
     )
 
@@ -486,21 +488,21 @@ class BumpPackages(CMD):
     given.
     """
 
-    arg_skip = Arg(
+    plugin_skip = Plugin(
         "-skip",
         help="Do not increment versions in these packages",
         nargs="+",
         default=[],
         required=False,
     )
-    arg_only = Arg(
+    plugin_only = Plugin(
         "-only",
         help="Only increment versions in these packages",
         nargs="+",
         default=[],
         required=False,
     )
-    arg_version = Arg("version", help="Version to increment by")
+    plugin_version = Plugin("version", help="Version to increment by")
 
     @staticmethod
     def bump_version(original, increment):