+++ service/http/tests/test_routes.py
@@ -20,7 +20,7 @@ from dffml.feature import DefFeature
 from dffml.source.memory import MemorySource, MemorySourceConfig
 from dffml.source.source import Sources
 from dffml.source.csv import CSVSourceConfig
-from dffml.util.cli.arg import parse_unknown
+from dffml.util.cli.plugin import parse_unknown
 from dffml.util.entrypoint import entrypoint
 from dffml.util.asynctestcase import AsyncTestCase
 from dffml.feature.feature import Feature, Features